First round of variable switching.
This demonstrates that our architecture can quite easily accommodated variable switching.  I've provided some test examples to illustrates how it works.

The examples show that variable switching in single-phase fluid flow (using either porepressure or saturation as the independent variable at each node) is probably a bad idea, at least for a van-genuchten capillary pressure.  By running the with_switching and without_switching examples, it's clear that the Jacobian in the with-switching is poorer which results in approximately 18-times more residual evaluations in this example.  (The example is infiltration of water into a 1D medium.)

Numerically there are two problems in this type of variable switching:

(1) In elements with some nodes being P and some nodes being S, the derivatives at the *quadpoints* that enter the Jacobian are not possible to calculate.  This is because they are not just the derivatives at the nodes weighted by the test functions, which is the case without variable switching.

(2) The values at the quadpoints are only properly calculable if all the nodes are P, or all are S.  Eg, if they are all P, then for linear-lagrange elements then P varies linearly over the element, whereas S doesn't (if S is a nonlinear function of P).  If all the nodes are P, then it's possible to use the capillary-pressure to calculate S at the quadpoints.  I do this.  However, if the elements are mixed type, then all you can say is that the nonlinear variable varies linearly, and then approximate P and S by a linear function, which is obviously wrong.

Having said this, perhaps a better formulation would be to use a *linear* capillary function.  That would fix (2) and probably (1) too.